First Improvments from Fork
Now uses node-config for startup from a config file (in my case a yaml)
https://github.com/lorenwest/node-config

Now works with a key-agent and also allows the text password access too via (awaiting testing)

modified the cli and remotecode class so that one can use the config file from the cli or direct from a remotecode instance.  Thus the defaults get set in remotecode class constructor and no longer in the cli.

Added a few other options
registry:  url of custom registry that will be used on install
key-agent: the environment variable holding the path to the agent socket,  there are a couple names tried by default

here is a sample config file
```
#commented keys indicate their default
options:
  target: /opt/relays
  #install: 'yarn'
  #registry:    # used for grabbing from custom registry first
  #source: .
  #start: 'nodemon .'  # pm2 could be used as well
  #verbose: false
  ssh:
    host: relays.kebler.net
    # port: 22
    username: sysadmin
    # need one of these three.  If none given will try to find the keyagent at some typical names
    #keyfilePath:
    #agent:  # the environment variable name of your ssh agent socket, defaults are SSH_AUTH_SOCK and SSH_AGENT_SOCK
    #password:  # for non keyfile access
    # keepaliveInterval: 500
    # readyTimeout: 2000
```